+++ Documentation/3.0/Samples/csharp/Raven.Documentation.Samples/ClientApi/Commands/Indexes/Put.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 
+using Raven.Abstractions.Data;
 using Raven.Abstractions.Indexing;
 using Raven.Client.Document;
 using Raven.Client.Indexes;
@@ -7,6 +8,7 @@ using Raven.Documentation.CodeSamples.Orders;
 
 namespace Raven.Documentation.Samples.ClientApi.Commands.Indexes
 {
+
 	public class Put
 	{
 		private interface IFoo
@@ -21,6 +23,10 @@ namespace Raven.Documentation.Samples.ClientApi.Commands.Indexes
 			string PutIndex<TDocument, TReduceResult>(string name, IndexDefinitionBuilder<TDocument, TReduceResult> indexDef);
 
 			string PutIndex<TDocument, TReduceResult>(string name, IndexDefinitionBuilder<TDocument, TReduceResult> indexDef, bool overwrite);
+			#endregion
+
+			#region put_indexes_3_0
+			string[] PutIndexes(IndexToAdd[] indexesToAdd);
 			#endregion
 		}
 
@@ -63,6 +69,53 @@ namespace Raven.Documentation.Samples.ClientApi.Commands.Indexes
 													Total = order.Lines.Sum(l => (l.Quantity * l.PricePerUnit) * (1 - l.Discount))
 												}
 							});
+				#endregion
+
+				#region put_indexes_3_1
+
+				store.DatabaseCommands.PutIndexes(new IndexToAdd[]
+				{
+					new IndexToAdd
+					{
+						Name = "Orders/Totals",
+						Priority = IndexingPriority.Normal,
+						Definition = new IndexDefinition
+						{
+							Map = @"from order in docs.Orders
+										select new 
+										{
+											order.Employee,
+											order.Company,
+											Total = order.Lines.Sum(l => (l.Quantity * l.PricePerUnit) * (1 - l.Discount))
+										}"
+						}
+					},
+					new IndexToAdd
+					{
+						Name = "Orders/ByCompany",
+						Priority = IndexingPriority.Normal,
+						Definition = new IndexDefinition
+						{
+							Map = @"from order in docs.Orders
+										select new 
+										{ 
+											order.Company, 
+											Count = 1, 
+											Total = order.Lines.Sum(l = >(l.Quantity * l.PricePerUnit) *  ( 1 - l.Discount))
+										}",
+
+							Reduce = @"from result in results
+										group result by result.Company into g
+										select new
+										{
+											Company = g.Key,
+											Count = g.Sum(x=>x.Count),
+											Total = g.Sum(x=>x.Total)
+										}"
+						}
+					}
+				});
+
 				#endregion
 			}
 		}