+++ Documentation/3.0/Raven.Documentation.Pages/start/whats-new.markdown
@@ -17,6 +17,7 @@
 
 - Added `AbstractScriptedIndexCreationTask`. More [here](../server/bundles/scripted-index-results#example-ii---abstractscriptedindexcreationtask),
 - Added `SetTransformerLock` command. More [here](../client-api/commands/transformers/how-to/change-transformer-lock-mode)
+- Added `PutIndexes` command. More [here](../client-api/commands/indexes/put#putindexes)
 
 {PANEL/}